ADM-363 move FormSelect from Common to CycleTime
